\name{ozib-package}
\alias{ozib-package}
\alias{ozib}
\docType{package}
\title{
\packageTitle{ozib}
}
\description{
\packageDescription{ozib}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{ozib}

Maintainer: \packageMaintainer{ozib}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ozib)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ozib}}
%\section{Documentation Index}{\packageIndices{ozib}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
